fix: RecipeIngredient 등록 및 수정 시퀀스 수정

+++ src/main/java/com/solution/recipetalk/service/recipe/ingredient/impl/ModifyRecipeIngredientServiceImpl.java
@@ -6,9 +6,11 @@ import com.solution.recipetalk.domain.recipe.ingredient.entity.RecipeIngredient;
 import com.solution.recipetalk.domain.recipe.ingredient.repository.RecipeIngredientRepository;
 import com.solution.recipetalk.domain.recipe.repository.RecipeRepository;
 import com.solution.recipetalk.dto.recipe.ingredient.RecipeIngredientRegisterDTO;
+import com.solution.recipetalk.exception.common.NotAuthorizedToModifyException;
 import com.solution.recipetalk.exception.ingredient.IngredientNotFoundException;
 import com.solution.recipetalk.exception.recipe.RecipeNotFoundException;
 import com.solution.recipetalk.service.recipe.ingredient.ModifyRecipeIngredientService;
+import com.solution.recipetalk.util.ContextHolder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -29,11 +31,16 @@ public class ModifyRecipeIngredientServiceImpl implements ModifyRecipeIngredient
     @Override
     public ResponseEntity<?> modifyRecipeIngredient(Long recipeId, List<RecipeIngredientRegisterDTO> dtos) {
         Recipe findRecipe = recipeRepository.findById(recipeId).orElseThrow(RecipeNotFoundException::new);
+
+        if(!recipeRepository.existsRecipeByBoardWriter_IdAndId(ContextHolder.getUserLoginId(), recipeId)){
+            throw new NotAuthorizedToModifyException();
+        }
+
         List<RecipeIngredient> recipeIngredientsByRecipeId = recipeIngredientRepository.findRecipeIngredientsByRecipeId(recipeId);
         recipeIngredientRepository.deleteAll(recipeIngredientsByRecipeId);
 
         List<RecipeIngredient> newRecipeIngredientList = dtos.stream().map(dto -> dto.toRecipeIngredientEntity(
-                        ingredientRepository.findById(dto.getIngredientId()).orElseThrow(IngredientNotFoundException::new),
+                        ingredientRepository.findById(dto.getIngredientId()).orElse(null),
                         findRecipe
                 )
         ).toList();
